services/api/requirements.txt
fastapi
uvicorn
psycopg2-binary
kafka-python
python-multipart
